==> TSTR_Pipeline/7_EvolutionaryOrigin_mmseqs.ipynb <==
## Evolutionary origin

Macaca - Mouse - Human

import os,re,glob
import pandas as pd
import numpy as np
from collections import Counter
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord
from Bio import SeqIO

species_dir = "/users/genomics/marta/TestisProject_SaraRazquin/with_TranscriptomeReconstruction"
users_dir = "/users/genomics/marta/TestisProject_SaraRazquin/with_TranscriptomeReconstruction/v47"

GENOMEDIR = "/genomics/users/marta/genomes"
species=["mouse","macaca","platypus","chicken","opossum"]

## annotation file
annotation="/users/genomics/marta/TestisProject_SaraRazquin/with_TranscriptomeReconstruction/v47/human/newReference_Resconstructed/gencode.v47.gffcompare.TestisLiverBrain.annotation.sorted.1transcript.sorted.NOchr.gtf"
transcript_gene=pd.read_csv("/users/genomics/marta/TestisProject_SaraRazquin/with_TranscriptomeReconstruction/v47/human/newReference_Resconstructed/transID_geneID_isoforms_selected.1to1.csv")

## evoDir
evoDir=os.path.join(users_dir,"EvolutionaryOrigin_MMseqs")

def translate_dna_to_protein(dna_seq):
    return str(Seq(dna_seq).translate())

# Function to save DF in fasta format
def create_seqrecord_PROT(row):
    return SeqRecord(Seq(row['protein']), id=row['header'][:-4], description="")
def create_seqrecord_DNA(row):
    return SeqRecord(Seq(row['seq']), id=row['header'], description="")

species_and_years_dict = {"specie" : ["human","macaca","mouse","opossum","platypus","chicken"],
                          "age" : [0,25,90,180,200,310]}
species_and_years = pd.DataFrame(data=species_and_years_dict, index=[1,2,3,4,5,6])
species_and_years


# ## get repre sequences - both DNA and PROTEINS
# for s in species:
#     print(s)
#     riboseq_dir = os.path.join(users_dir,s,"RiboSeq/RiboQC_RiboNovel")

#     ## get all candidates
#     print("Reading all the candidate ORF sequences")
#     df = pd.read_csv(os.path.join(riboseq_dir,"Annotation/candidateORF.fa"), header=None, sep="\t")
#     ## convert to df
#     candidates_fasta = pd.DataFrame({'header':df[0].iloc[::2].values, 'seq':df[0].iloc[1::2].values})
#     candidates_fasta['header'] = candidates_fasta['header'].str[1:]
#     candidates_fasta = candidates_fasta[~candidates_fasta['header'].str.contains("iORF")]

#     all_repre_candidates = pd.DataFrame()
#     for sample in os.listdir(os.path.join(riboseq_dir,"RibORF")):
#         if sample.endswith("r1"):
#             print(sample)

#             repre = pd.read_csv(os.path.join(riboseq_dir,"RibORF",sample,"repre.valid.pred.pvalue.parameters.txt"), sep="\t")
#             repre_fasta = candidates_fasta[candidates_fasta['header'].isin(repre.orfID.values.tolist())]
#             repre_fasta['header'] = s + "_" +repre_fasta['header']

#             ## keep only the coding and the non-coding
#             repre_fasta = repre_fasta[repre_fasta['header'].str.contains("canonical")| repre_fasta['header'].str.contains("noncoding")]

#             repre_fasta['protein'] = repre_fasta['seq'].apply(translate_dna_to_protein)
#             all_repre_candidates = pd.concat([all_repre_candidates, repre_fasta])

#             ## translate to protein
#             print("Translating to protein")
#             out_proteins = os.path.join(riboseq_dir,"RibORF",sample,"repre.valid.pred.pvalue.parameters.PROTEIN.fa")
#             seq_records_proteins = repre_fasta.apply(create_seqrecord_PROT, axis=1).tolist()
#             SeqIO.write(seq_records_proteins, out_proteins, "fasta")

#             ## keep the dna sequences as well
#             print("Saving the original DNA ORF")
#             out_dna = os.path.join(riboseq_dir,"RibORF",sample,"repre.valid.pred.pvalue.parameters.DNA.fa")
#             seq_records_dna = repre_fasta.apply(create_seqrecord_DNA, axis=1).tolist()
#             SeqIO.write(seq_records_dna, out_dna, "fasta")

#             ## get only the nonredundant
#             print("Getting the non-redundant")
#             all_repre_candidates.drop_duplicates(inplace=True)
#             out_nonredundant_proteins = os.path.join(riboseq_dir,"RibORF")+"/"+s+"_repre.valid.pred.pvalue.parameters.PROTEIN.noredunant.fa"
#             seq_records_proteins = all_repre_candidates.apply(create_seqrecord_PROT, axis=1).tolist()
#             SeqIO.write(seq_records_proteins, out_nonredundant_proteins, "fasta")

## MMseqs2

Against all expressed in mouse, macaca, platypus, opossum in any tissue (non-redunant set)

#### TESTIS-EXPRESSED

## Testis-expressed ORFs human
TestisExpressed_translatedINtestis = pd.read_csv("/projects_eg/projects/marta/TestisRestricted_Microproteins_TSA/with_TranscriptomeReconstruction/Multimap_altORFs/Q1_TestisORFs/human/ribORF_humanTestis_in1.csv")
TestisExpressed_translatedINtestis

def create_seqrecord_PROThuman(row):
    return SeqRecord(Seq(row['ORFpep']), id=row['orfID'], description="")

out_nonredundant_proteins = "/projects_eg/projects/marta/TestisRestricted_Microproteins_TSA/with_TranscriptomeReconstruction/Multimap_altORFs/Q1_TestisORFs/human/ribORF_humanTestis_in1.fa"
seq_records_proteins = TestisExpressed_translatedINtestis.apply(create_seqrecord_PROThuman, axis=1).tolist()
SeqIO.write(seq_records_proteins, out_nonredundant_proteins, "fasta")

Creating database of human (expressed in testis)

%%bash -s "$evoDir" "$users_dir"

# mkdir $1/MMseqsDB

module load Miniconda3/4.9.2
source /soft/system/software/Miniconda3/4.9.2/bin/activate MMseqs2

mkdir -p $1/MMseqsDB

mmseqs createdb /projects_eg/projects/marta/TestisRestricted_Microproteins_TSA/with_TranscriptomeReconstruction/Multimap_altORFs/Q1_TestisORFs/human/ribORF_humanTestis_in1.fa $1/MMseqsDB/ribORF_humanTestis_in1_DB


Creating a joint database with all the specie, everything translated, no matter the tissue

%%bash -s "$evoDir" "$species_dir"

module load Miniconda3/4.9.2
source /soft/system/software/Miniconda3/4.9.2/bin/activate MMseqs2

# species="platypus mouse macaca chicken"
# for s in $species; do
#     echo $s

#     ### create DB
#     mmseqs createdb $2/$s/RiboSeq/RiboQC_RiboNovel/RibORF/repre.valid.pred.pvalue.parameters.PROTEIN.noredunant.fa $1/MMseqsDB/${s}_noredunantProts_DB

# done

mmseqs createdb $2/*/RiboSeq/RiboQC_RiboNovel*/RibORF/*_repre.valid.pred.pvalue.parameters.PROTEIN.noredunant.fa $1/MMseqsDB/speciesDB



Creating a joint database with all the specie AND human for clustering, everything translated for the species, no matter the tissue, but only translated in 1 for testis

%%bash -s "$evoDir" "$species_dir"

module load Miniconda3/4.9.2
source /soft/system/software/Miniconda3/4.9.2/bin/activate MMseqs2

mmseqs createdb $2/*/RiboSeq/RiboQC_RiboNovel*/RibORF/*_repre.valid.pred.pvalue.parameters.PROTEIN.noredunant.fa /projects_eg/projects/marta/TestisRestricted_Microproteins_TSA/with_TranscriptomeReconstruction/Multimap_altORFs/Q1_TestisORFs/human/ribORF_humanTestis_in1.fa $1/MMseqsDB/species_and_humanDB



Searching

# %%bash -s "$evoDir" "$users_dir"

# module load Miniconda3/4.9.2
# source /soft/system/software/Miniconda3/4.9.2/bin/activate MMseqs2
# mkdir -p $1/Searching

# rm $1/Searching/resultDB*
# ## The alignment consists of two steps the prefilter and alignment. To run the search, type:
# mmseqs search -a -s 6 $1/MMseqsDB/ribORF_humanTestis_in1_DB $1/MMseqsDB/speciesDB $1/Searching/resultDB $1/tmp

# # Then, convert the result database into a BLAST tab formatted file (option -m 8 in legacy blast, -outfmt 6 in blast+):
# # Format alignment output	query,target,fident,alnlen,mismatch,gapopen,qstart,qend,tstart,tend,evalue,bits

# mmseqs convertalis $1/MMseqsDB/ribORF_humanTestis_in1_DB $1/MMseqsDB/speciesDB $1/Searching/resultDB $1/Searching/resultDB.m8

Clustering

%%bash -s "$evoDir" "$users_dir"

module load Miniconda3/4.9.2
source /soft/system/software/Miniconda3/4.9.2/bin/activate MMseqs2

mkdir -p $1/Clustering

rm -r $1/Clustering/species_and_human_cluDB*

## cluster all species and human
mmseqs cluster -c 0.5 --min-seq-id 0.5 $1/MMseqsDB/species_and_humanDB $1/Clustering/species_and_human_cluDB $1/tmp
mmseqs createtsv $1/MMseqsDB/species_and_humanDB $1/MMseqsDB/species_and_humanDB $1/Clustering/species_and_human_cluDB $1/Clustering/species_and_human_cluDB.tsv

## Analysis

clustering = pd.read_csv(os.path.join(evoDir,"Clustering/species_and_human_cluDB.tsv"), header=None, sep="\t")
clustering = clustering[clustering[0].str.startswith("ENST") | clustering[0].str.startswith("TCONS")]
print("all human")
print(len(clustering))
clustering['transcript_id'] = clustering[0].str.split(":", expand=True)[0]
clustering['transcript_id'] = clustering["transcript_id"].str.split(".", expand=True)[0]

clustering = clustering.merge(transcript_gene, on="transcript_id")
print("all human after annotation")
print(len(clustering))
small_clust = clustering[["gene_name","gene_type",0,1]]
small_clust.groupby("gene_type").count()


small_clust

## non-coding
noncoding = clustering[clustering[0].str.contains("noncoding")]
noncoding.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_speciesDB.tsv"), header=None, index=None, sep="\t")
print("noncoding all")
print(len(noncoding))
print(len(set(noncoding[0].values.tolist())))

## altORFs
altORFs = clustering[clustering[0].str.contains("ORF")]
altORFs.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_speciesDB.tsv"), header=None, index=None, sep="\t")
print("altORFs all")
print(len(altORFs))
print(len(set(altORFs[0].values.tolist())))

## canonical
canonical = clustering[clustering[0].str.contains("canonical")]
print("Canonical all")
print(len(canonical))
print(len(set(canonical[0].values.tolist())))
canonical.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_speciesDB.tsv"), header=None, index=None, sep="\t")

#### NONCODING

# if column 1 starts with a specie, it is conserved
noncoding_conserved = noncoding[~noncoding[1].str.startswith("ENST")]
noncoding_conserved = noncoding_conserved[~noncoding_conserved[1].str.startswith("TCONS")]
noncoding_conserved.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_Conserved.tsv"), header=None, index=None, sep="\t")
print("Conserved noncoding human ORFs")
print(len(set(noncoding_conserved[0].values.tolist())))
noncoding_conserved = noncoding_conserved[['gene_name','gene_type',0,1]]
noncoding_conserved.columns = ['gene_name','gene_type','orfID',1]

# if column 1 starts with a human gene, it is not conserved
noncoding_NonConserved = noncoding[noncoding[1].str.startswith("ENST") | noncoding[1].str.startswith("TCONS")]
noncoding_NonConserved.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_NonConserved.tsv"), header=None, index=None, sep="\t")
print("Non-Conserved noncoding human ORFs")
print(len(set(noncoding_NonConserved[0].values.tolist())))
noncoding_NonConserved = noncoding_NonConserved[['gene_name','gene_type',0,1]]
noncoding_NonConserved.columns = ['gene_name','gene_type','orfID',1]
noncoding_NonConserved

## conserved non-coding
noncoding_conserved_to_matrix = noncoding_conserved
print(len(set(noncoding_conserved_to_matrix.orfID.values.tolist())))
# Assuming noncoding_conserved_to_matrix is already defined
noncoding_conserved_to_matrix['specie'] = noncoding_conserved_to_matrix[1].str.split("_", expand=True)[0]
noncoding_conserved_to_matrix = noncoding_conserved_to_matrix[['orfID', 'specie']]
noncoding_conserved_to_matrix.groupby('specie').count()
noncoding_conserved_to_matrix['presence'] = 1
## control
print(len(set(noncoding_conserved_to_matrix.orfID.values.tolist())))

# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
noncoding_conserved_matrix = noncoding_conserved_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')

# Fill NaN values with 0
noncoding_conserved_matrix = noncoding_conserved_matrix.fillna(0)

# Reset index so 'orfID' becomes a column instead of an index
noncoding_conserved_matrix = noncoding_conserved_matrix.reset_index()

# Reorder columns for the final matrix
noncoding_conserved_matrix_ordered = noncoding_conserved_matrix[["orfID", "macaca", "mouse", "opossum", "platypus", "chicken"]]

# Calculate sum across species columns to identify how many species are present
noncoding_conserved_matrix_ordered['sum'] = noncoding_conserved_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
noncoding_conserved_matrix_ordered = noncoding_conserved_matrix_ordered.sort_values(by="sum", ascending=False)
print(len(set(noncoding_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
noncoding_conserved_matrix_ordered =  noncoding_conserved_matrix_ordered.merge(noncoding_conserved[['gene_name','gene_type','orfID']], on="orfID")
noncoding_conserved_matrix_ordered.drop_duplicates(inplace=True)

noncoding_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(noncoding_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
noncoding_conserved_matrix_ordered =  noncoding_conserved_matrix_ordered.merge(noncoding_conserved[['gene_name','gene_type','orfID']], on="orfID")
noncoding_conserved_matrix_ordered.drop_duplicates(inplace=True)

noncoding_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(noncoding_conserved_matrix_ordered.orfID.values.tolist())))


### all noncoding
noncoding_conserved_touse = noncoding_conserved
noncoding_conserved_touse['specie'] = noncoding_conserved_touse[1].str.split("_", expand=True)[0]
# noncoding_NonConserved.drop("specie",axis=1, inplace=True)
noncoding_to_matrix_raw = pd.concat([noncoding_conserved, noncoding_NonConserved])
noncoding_to_matrix = noncoding_to_matrix_raw[['orfID', 'specie']]
noncoding_to_matrix = noncoding_to_matrix.fillna('human').drop_duplicates()
noncoding_to_matrix['presence'] = 1
print(noncoding_to_matrix.groupby("specie").count())
noncoding_to_matrix

# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
noncoding_matrix = noncoding_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')
# Fill NaN values with 0
noncoding_matrix = noncoding_matrix.fillna(0)
# Reset index so 'orfID' becomes a column instead of an index
noncoding_matrix = noncoding_matrix.reset_index()

# Rename columns as necessary
noncoding_matrix.columns = ["orfID", "chicken", "human", "macaca", "mouse", "opossum", "platypus"]
print(len(noncoding_matrix))
# Reorder columns for the final matrix
noncoding_matrix_ordered = noncoding_matrix[["orfID", "human","macaca", "mouse", "opossum", "platypus", "chicken"]]

# # Calculate sum across species columns to identify how many species are present
noncoding_matrix_ordered['sum'] = noncoding_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
noncoding_matrix_ordered = noncoding_matrix_ordered.sort_values(by="sum", ascending=False)

# Display the result
noncoding_matrix_ordered = noncoding_matrix_ordered.merge(noncoding_to_matrix_raw[['gene_name','gene_type','orfID']], on="orfID")
noncoding_matrix_ordered.drop_duplicates(inplace=True)

noncoding_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_matrix.tsv"), sep="\t", index=None)
print(len(set(noncoding_matrix_ordered.orfID.values.tolist())))
noncoding_matrix_ordered


# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
noncoding_matrix = noncoding_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')
noncoding_matrix
# Fill NaN values with 0
noncoding_matrix = noncoding_matrix.fillna(0)

# Reset index so 'orfID' becomes a column instead of an index
noncoding_matrix = noncoding_matrix.reset_index()

# Rename columns as necessary
noncoding_matrix.columns = ["orfID", "chicken", "human", "macaca", "mouse", "opossum", "platypus"]

# Reorder columns for the final matrix
noncoding_matrix_ordered = noncoding_matrix[["orfID", "human","macaca", "mouse", "opossum", "platypus", "chicken"]]

# Calculate sum across species columns to identify how many species are present
noncoding_matrix_ordered['sum'] = noncoding_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
noncoding_matrix_ordered = noncoding_matrix_ordered.sort_values(by="sum", ascending=False)

# Display the result
noncoding_matrix_ordered =  noncoding_matrix_ordered.merge(noncoding_to_matrix_raw[['gene_name','gene_type','orfID']], on="orfID")
noncoding_matrix_ordered.drop_duplicates(inplace=True)

noncoding_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_matrix.tsv"), sep="\t", index=None)
print(len(set(noncoding_matrix_ordered.orfID.values.tolist())))

### alternative ORFs

# if column 1 starts with a specie, it is conserved
altORFs_conserved = altORFs[~altORFs[1].str.startswith("ENST")]
altORFs_conserved.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_Conserved.tsv"), header=None, index=None, sep="\t")
print("Conserved altORFs human ORFs")
print(len(set(altORFs_conserved[0].values.tolist())))
altORFs_conserved = altORFs_conserved[['gene_name','gene_type',0,1]]
altORFs_conserved.columns = ['gene_name','gene_type','orfID',1]

# if column 1 starts with a human gene, it is not conserved
altORFs_NonConserved = altORFs[altORFs[1].str.startswith("ENST")]
altORFs_NonConserved.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_NonConserved.tsv"), header=None, index=None, sep="\t")
print("Not-conserved altORFs human ORFs")
print(len(set(altORFs_NonConserved[0].values.tolist())))
altORFs_NonConserved = altORFs_NonConserved[['gene_name','gene_type',0,1]]
altORFs_NonConserved.columns = ['gene_name','gene_type','orfID',1]


## conserved altORFs
altORFs_conserved_to_matrix = altORFs_conserved
print(len(set(altORFs_conserved_to_matrix.orfID.values.tolist())))
# Assuming noncoding_conserved_to_matrix is already defined
altORFs_conserved_to_matrix['specie'] = altORFs_conserved_to_matrix[1].str.split("_", expand=True)[0]
altORFs_conserved_to_matrix = altORFs_conserved_to_matrix[['orfID', 'specie']]
altORFs_conserved_to_matrix.groupby('specie').count()
altORFs_conserved_to_matrix['presence'] = 1
## control
print(len(set(altORFs_conserved_to_matrix.orfID.values.tolist())))

# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
altORFs_conserved_matrix = altORFs_conserved_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')

# Fill NaN values with 0
altORFs_conserved_matrix = altORFs_conserved_matrix.fillna(0)

# Reset index so 'orfID' becomes a column instead of an index
altORFs_conserved_matrix = altORFs_conserved_matrix.reset_index()

# Reorder columns for the final matrix
altORFs_conserved_matrix_ordered = altORFs_conserved_matrix[["orfID", "macaca", "mouse", "opossum", "platypus", "chicken"]]

# Calculate sum across species columns to identify how many species are present
altORFs_conserved_matrix_ordered['sum'] = altORFs_conserved_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
altORFs_conserved_matrix_ordered = altORFs_conserved_matrix_ordered.sort_values(by="sum", ascending=False)
print(len(set(altORFs_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
altORFs_conserved_matrix_ordered =  altORFs_conserved_matrix_ordered.merge(altORFs_conserved[['gene_name','gene_type','orfID']], on="orfID")
altORFs_conserved_matrix_ordered.drop_duplicates(inplace=True)

altORFs_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(altORFs_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
altORFs_conserved_matrix_ordered =  altORFs_conserved_matrix_ordered.merge(altORFs_conserved[['gene_name','gene_type','orfID']], on="orfID")
altORFs_conserved_matrix_ordered.drop_duplicates(inplace=True)

altORFs_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(altORFs_conserved_matrix_ordered.orfID.values.tolist())))


### all altORFs
altORFs_conserved_touse = altORFs_conserved
altORFs_conserved_touse['specie'] = altORFs_conserved_touse[1].str.split("_", expand=True)[0]
altORFs_to_matrix_raw = pd.concat([altORFs_conserved, altORFs_NonConserved])
altORFs_to_matrix = altORFs_to_matrix_raw[['orfID', 'specie']]
altORFs_to_matrix = altORFs_to_matrix.fillna('human').drop_duplicates()
altORFs_to_matrix['presence'] = 1
altORFs_to_matrix = altORFs_to_matrix.replace("TCONS", "human")
print(altORFs_to_matrix.groupby("specie").count())


# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
altORFs_matrix = altORFs_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')
# Fill NaN values with 0
altORFs_matrix = altORFs_matrix.fillna(0)
# Reset index so 'orfID' becomes a column instead of an index
altORFs_matrix = altORFs_matrix.reset_index()
altORFs_matrix
# Rename columns as necessary
altORFs_matrix.columns = ["orfID", "chicken", "human", "macaca", "mouse", "opossum", "platypus"]
print(len(altORFs_matrix))
# Reorder columns for the final matrix
altORFs_matrix_ordered = altORFs_matrix[["orfID", "human","macaca", "mouse", "opossum", "platypus", "chicken"]]

# # Calculate sum across species columns to identify how many species are present
altORFs_matrix_ordered['sum'] = altORFs_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
altORFs_matrix_ordered = altORFs_matrix_ordered.sort_values(by="sum", ascending=False)

# Display the result
altORFs_matrix_ordered = altORFs_matrix_ordered.merge(altORFs_to_matrix_raw[['gene_name','gene_type','orfID']], on="orfID")
altORFs_matrix_ordered.drop_duplicates(inplace=True)

altORFs_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_matrix.tsv"), sep="\t", index=None)
print(len(set(altORFs_matrix_ordered.orfID.values.tolist())))
altORFs_matrix_ordered


### CANONICAL

# if column 1 starts with a specie, it is conserved
canonical_conserved = canonical[~canonical[1].str.startswith("ENST")]
canonical_conserved.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_Conserved.tsv"), header=None, index=None, sep="\t")
print("Conserved canonical human ORFs")
print(len(set(canonical_conserved[0].values.tolist())))
canonical_conserved = canonical_conserved[['gene_name','gene_type',0,1]]
canonical_conserved.columns = ['gene_name','gene_type','orfID',1]

# if column 1 starts with a human gene, it is not conserved
canonical_NonConserved = canonical[canonical[1].str.startswith("ENST")]
canonical_NonConserved.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_NonConserved.tsv"), header=None, index=None, sep="\t")
print("AltORFs noncoding human ORFs")
print(len(set(canonical_NonConserved[0].values.tolist())))
canonical_NonConserved = canonical_NonConserved[['gene_name','gene_type',0,1]]
canonical_NonConserved.columns = ['gene_name','gene_type','orfID',1]


## conserved canonical
canonical_conserved_to_matrix = canonical_conserved
print(len(set(canonical_conserved_to_matrix.orfID.values.tolist())))
# Assuming noncoding_conserved_to_matrix is already defined
canonical_conserved_to_matrix['specie'] = canonical_conserved_to_matrix[1].str.split("_", expand=True)[0]
canonical_conserved_to_matrix = canonical_conserved_to_matrix[['orfID', 'specie']]
canonical_conserved_to_matrix.groupby('specie').count()
canonical_conserved_to_matrix['presence'] = 1
## control
print(len(set(canonical_conserved_to_matrix.orfID.values.tolist())))

# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
canonical_conserved_matrix = canonical_conserved_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')

# Fill NaN values with 0
canonical_conserved_matrix = canonical_conserved_matrix.fillna(0)

# Reset index so 'orfID' becomes a column instead of an index
canonical_conserved_matrix = canonical_conserved_matrix.reset_index()

# Reorder columns for the final matrix
canonical_conserved_matrix_ordered = canonical_conserved_matrix[["orfID", "macaca", "mouse", "opossum", "platypus", "chicken"]]

# Calculate sum across species columns to identify how many species are present
canonical_conserved_matrix_ordered['sum'] = canonical_conserved_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
canonical_conserved_matrix_ordered = canonical_conserved_matrix_ordered.sort_values(by="sum", ascending=False)
print(len(set(canonical_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
canonical_conserved_matrix_ordered =  canonical_conserved_matrix_ordered.merge(canonical_conserved[['gene_name','gene_type','orfID']], on="orfID")
canonical_conserved_matrix_ordered.drop_duplicates(inplace=True)

canonical_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(canonical_conserved_matrix_ordered.orfID.values.tolist())))

# Display the result
canonical_conserved_matrix_ordered =  canonical_conserved_matrix_ordered.merge(canonical_conserved[['gene_name','gene_type','orfID']], on="orfID")
canonical_conserved_matrix_ordered.drop_duplicates(inplace=True)

canonical_conserved_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_Conserved_matrix.tsv"), sep="\t", index=None)
print(len(set(canonical_conserved_matrix_ordered.orfID.values.tolist())))


### all canonical
canonical_conserved_touse = canonical_conserved
canonical_conserved_touse['specie'] = canonical_conserved_touse[1].str.split("_", expand=True)[0]
canonical_to_matrix_raw = pd.concat([canonical_conserved, canonical_NonConserved])
canonical_to_matrix = canonical_to_matrix_raw[['orfID', 'specie']]
canonical_to_matrix = canonical_to_matrix.fillna('human').drop_duplicates()
canonical_to_matrix['presence'] = 1
canonical_to_matrix = canonical_to_matrix.replace("TCONS", "human")
print(canonical_to_matrix.groupby("specie").count())


# Pivoting the data without creating an 'id', and keeping 'orfID' as the unique identifier
canonical_matrix = canonical_to_matrix.pivot_table(index='orfID', columns='specie', values='presence', aggfunc='max')
# Fill NaN values with 0
canonical_matrix = canonical_matrix.fillna(0)
# Reset index so 'orfID' becomes a column instead of an index
canonical_matrix = canonical_matrix.reset_index()
canonical_matrix
# Rename columns as necessary
canonical_matrix.columns = ["orfID", "chicken", "human", "macaca", "mouse", "opossum", "platypus"]
print(len(canonical_matrix))
# Reorder columns for the final matrix
canonical_matrix_ordered = canonical_matrix[["orfID", "human","macaca", "mouse", "opossum", "platypus", "chicken"]]

# # Calculate sum across species columns to identify how many species are present
canonical_matrix_ordered['sum'] = canonical_matrix_ordered.select_dtypes(include='number').sum(axis=1)

# Sort by the sum column to order by the number of species present
canonical_matrix_ordered = canonical_matrix_ordered.sort_values(by="sum", ascending=False)

# Display the result
canonical_matrix_ordered = canonical_matrix_ordered.merge(canonical_to_matrix_raw[['gene_name','gene_type','orfID']], on="orfID")
canonical_matrix_ordered.drop_duplicates(inplace=True)

canonical_matrix_ordered.to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_matrix.tsv"), sep="\t", index=None)
print(len(set(canonical_matrix_ordered.orfID.values.tolist())))
canonical_matrix_ordered

Find the oldest specie containg the ORF translated

species_ordered = ['chicken','platypus','opossum','mouse','macaca','human']
def find_last_species(row):
    for sp in species_ordered:
        if row[sp] == 1.0:  # Check if the value is 1.0
            return sp
    return None  # Return None if no species has a 1

# Apply the function to each row to create a new column 'oldest_species'
canonical_matrix_ordered['oldest_species'] = canonical_matrix_ordered.apply(find_last_species, axis=1)
canonical_matrix_ordered = species_and_years.merge(canonical_matrix_ordered, left_on=["specie"], right_on="oldest_species")
canonical_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']].to_csv(os.path.join(evoDir,"Clustering/canonicalHuman_matrix_oldestAge.tsv"), sep="\t", index=None)

noncoding_matrix_ordered['oldest_species'] = noncoding_matrix_ordered.apply(find_last_species, axis=1)
noncoding_matrix_ordered = species_and_years.merge(noncoding_matrix_ordered, left_on=["specie"], right_on="oldest_species")
noncoding_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']].to_csv(os.path.join(evoDir,"Clustering/noncodingHuman_matrix_oldestAge.tsv"), sep="\t", index=None)


altORFs_matrix_ordered['oldest_species'] = altORFs_matrix_ordered.apply(find_last_species, axis=1)
altORFs_matrix_ordered = species_and_years.merge(altORFs_matrix_ordered, left_on=["specie"], right_on="oldest_species")
altORFs_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']].to_csv(os.path.join(evoDir,"Clustering/altORFsHuman_matrix_oldestAge.tsv"), sep="\t", index=None)


complete_matrix_age = pd.concat([noncoding_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']],altORFs_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']], canonical_matrix_ordered[['orfID','gene_name','gene_type','oldest_species','age']]])
complete_matrix_age.to_csv(os.path.join(evoDir,"Clustering/allHuman_matrix_oldestAge.tsv"), sep="\t", index=None)
