==> TSTR_Pipeline/11_predictions_immunopeptidomics.ipynb <==
import os, re
import pandas as pd

IPdir = "/users/genomics/marta/241205_Immunopeptidomics_HCC"
PREDdir = "/users/genomics/marta/TestisProject_SaraRazquin/with_TranscriptomeReconstruction/v47/cancers/log2ratio3x/cancertypes/onlyStep1/netMHCpan/FrequentHLA-I"

immunopeptidomics_results = pd.read_csv(os.path.join(IPdir,"results/merged/merged_results_HCC_LUNG_Melanoma.csv"))

print(len(set(immunopeptidomics_results.sequence.values.tolist())),"unique hits in TSTR genes with immunopeptidomics data (HCC + lung + melanoma, FDR5%)")
immunopeptidomics_results


## WB / SB

## exact hit
for file in os.listdir(PREDdir):
    if file.endswith("WBSB.out"):
        print(file.split(".")[0])
        WB = pd.read_csv(os.path.join(PREDdir,file), header=None, sep='\s+')
        WB.drop([0,3,9,4,5,6,7,8,16,17], axis=1, inplace=True)
        WB.columns = ["MHC","Peptide","Identity","Score_EL","%Rank_EL","Score_BA","%Rank_BA","Aff(nM)"]
        print(len(set(WB.Peptide.values.tolist())),"predicted binders in ",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        
        # Filter immunopeptidomics_results to keep rows where sequence is the same as WB peptide
        WB_IP = WB[['Peptide']].merge(immunopeptidomics_results, left_on="Peptide", right_on="sequence")
        print(len(set(WB_IP.Peptide.values.tolist())),"overlapping of hits",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        # print(WB_IP.drop_duplicates())

## WB contained in IP hit
for file in os.listdir(PREDdir):
    if file.endswith("WBSB.out"):
        print(file.split(".")[0])
        WB = pd.read_csv(os.path.join(PREDdir,file), header=None, sep='\s+')
        WB.drop([0,3,9,4,5,6,7,8,16,17], axis=1, inplace=True)
        WB.columns = ["MHC","Peptide","Identity","Score_EL","%Rank_EL","Score_BA","%Rank_BA","Aff(nM)"]
        print(len(set(WB.Peptide.values.tolist())),"predicted binders in ",file.split(".")[0],"TSTR genes for any common HLA-I allele")

        # Filter immunopeptidomics_results to keep rows where sequence is the same as or contains a WB peptide
        WB_peptides = WB.Peptide.values.tolist()
        WB_IP = immunopeptidomics_results[
                    immunopeptidomics_results['sequence'].apply(lambda seq: any(peptide in seq for peptide in WB_peptides))]     
           
        print(len(set(WB_IP.sequence.values.tolist())),"overlapping of hits",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        # print(WB_IP.drop_duplicates())

## SB

## exact hit
for file in os.listdir(PREDdir):
    if file.endswith(".SB.out"):
        print(file.split(".")[0])
        SB = pd.read_csv(os.path.join(PREDdir,file), header=None, sep='\s+')
        SB.drop([0,3,9,4,5,6,7,8,16,17], axis=1, inplace=True)
        SB.columns = ["MHC","Peptide","Identity","Score_EL","%Rank_EL","Score_BA","%Rank_BA","Aff(nM)"]
        print(len(set(SB.Peptide.values.tolist())),"predicted strong binders in ",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        
        # Filter immunopeptidomics_results to keep rows where sequence is the same as WB peptide
        SB_IP = SB[['Peptide']].merge(immunopeptidomics_results, left_on="Peptide", right_on="sequence")
        print(len(set(SB_IP.Peptide.values.tolist())),"overlapping of hits",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        # print(SB_IP.drop_duplicates())

## WB contained in IP hit
for file in os.listdir(PREDdir):
    if file.endswith(".SB.out"):
        print(file.split(".")[0])
        SB = pd.read_csv(os.path.join(PREDdir,file), header=None, sep='\s+')
        SB.drop([0,3,9,4,5,6,7,8,16,17], axis=1, inplace=True)
        SB.columns = ["MHC","Peptide","Identity","Score_EL","%Rank_EL","Score_BA","%Rank_BA","Aff(nM)"]
        print(len(set(SB.Peptide.values.tolist())),"predicted strong binders in ",file.split(".")[0],"TSTR genes for any common HLA-I allele")

        # Filter immunopeptidomics_results to keep rows where sequence is the same as or contains a WB peptide
        SB_peptides = SB.Peptide.values.tolist()
        SB_IP = immunopeptidomics_results[
                    immunopeptidomics_results['sequence'].apply(lambda seq: any(peptide in seq for peptide in SB_peptides))]     
           
        print(len(set(SB_IP.sequence.values.tolist())),"overlapping of hits",file.split(".")[0],"TSTR genes for any common HLA-I allele")
        # print(SB_IP.drop_duplicates())
